backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
flask
flask-cors
werkzeug

# ─── OpenRouter / OpenAI SDK ────────────────────────────────────────────────────────
openai

# ─── Document reading ─────────────────────────────────────────────────────────
pdfplumber
pytesseract
Pillow

# ─── RAG stack ────────────────────────────────────────────────────────────────
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers

# ─── Structured data ──────────────────────────────────────────────────────────
pandas

# ─── Validation ───────────────────────────────────────────────────────────────
pydantic

# ─── Configuration ────────────────────────────────────────────────────────────
python-dotenv

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest
pytest-asyncio

# ─── Transitive deps ──────────────────────────────────────────────────────────
httpx
anyio
mcp[cli]
gunicorn